requirements.txt
absl-py==2.3.0
altair==5.5.0
appdirs==1.4.4
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.1
frozendict==2.4.6
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.2
grpcio==1.72.1
h5py==3.13.0
html5lib==1.1
idna==3.10
importlib-metadata==6.11.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.12.0
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==3.3.5
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
narwhals==1.41.0
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==23.2
pandas==1.5.3
pandas_ta==0.3.14b0
Pillow==9.5.0
plotly==6.1.2
protobuf==4.25.8
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==310
PyYAML==6.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.3.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soupsieve==2.7
SQLAlchemy==2.0.18
streamlit==1.23.1
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.14.1
yfinance==0.2.26
zipp==3.22.0
